feat!: sdk regeneration 2026-04-24 (#696)
## Summary

- regenerate the Python SDK with the latest Fern toolchain
- add generated Agent History support and fix usage breakdown `tags` to
`list[str]`
- re-apply the required websocket client patches
- restore History websocket parsing in `agent/v1`
- align CI and docs with Python `3.10+`

## Validation

- `uv run ruff check src`
- `uv run mypy src`
- `uv run pytest -q` (`172 passed, 4 skipped`)

---------

+++ src/deepgram/agent/v1/types/agent_v1settings_audio_input.py
@@ -18,7 +18,7 @@ class AgentV1SettingsAudioInput(UncheckedBaseModel):
     Audio encoding format
     """
 
-    sample_rate: float = pydantic.Field()
+    sample_rate: int = pydantic.Field()
     """
     Sample rate in Hz. Common values are 16000, 24000, 44100, 48000
     """
