+++ src/deepgram/listen/v1/requests/listen_v1metadata.py
@@ -36,7 +36,7 @@ class ListenV1MetadataParams(typing_extensions.TypedDict):
     The duration
     """
 
-    channels: float
+    channels: int
     """
     The channels
     """
