+++ src/deepgram/listen/v1/requests/listen_v1utterance_end.py
@@ -11,7 +11,7 @@ class ListenV1UtteranceEndParams(typing_extensions.TypedDict):
     Message type identifier
     """
 
-    channel: typing.Sequence[float]
+    channel: typing.Sequence[int]
     """
     The channel
     """
