+++ src/deepgram/listen/v2/types/listen_v2fatal_error.py
@@ -13,7 +13,7 @@ class ListenV2FatalError(UncheckedBaseModel):
     Message type identifier
     """
 
-    sequence_id: float = pydantic.Field()
+    sequence_id: int = pydantic.Field()
     """
     Starts at `0` and increments for each message the server sends
     to the client.  This includes messages of other types, like
