+++ src/deepgram/requests/listen_v1response_results_channels_item_alternatives_item_paragraphs_paragraphs_item.py
@@ -12,7 +12,7 @@ class ListenV1ResponseResultsChannelsItemAlternativesItemParagraphsParagraphsIte
     sentences: typing_extensions.NotRequired[
         typing.Sequence[ListenV1ResponseResultsChannelsItemAlternativesItemParagraphsParagraphsItemSentencesItemParams]
     ]
-    speaker: typing_extensions.NotRequired[float]
-    num_words: typing_extensions.NotRequired[float]
+    speaker: typing_extensions.NotRequired[int]
+    num_words: typing_extensions.NotRequired[int]
     start: typing_extensions.NotRequired[float]
     end: typing_extensions.NotRequired[float]
